pyproject.toml
[tool.poetry]
name = "marvelpipeline"
version = "0.1.0"
description = "The pipline used for the MARVEl telescope"
authors = ["Dries Seynaeve <dries.seynaeve@kuleuven.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
h5py = "^3.6"
numpy = "^1.22.3"
pandas = "^1.4.1"
PyYAML = "^6.0.1"
astropy = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
